+++ src/controllers/LGUTimetable.js
@@ -1,0 +1,13 @@
+const { Timetable } = require('../models');
+
+class LGUTimetable {
+  static async get(req, res) {
+    const { session, semester, program, section } = req.data;
+
+    const timetable = new Timetable(session, semester, program, section);
+    res.send(await timetable.getAll());
+  }
+  static async getByDay(req, res) {}
+}
+
+module.exports = { LGUTimetable };
