+++ src/routes/LGUTimetable.js
@@ -1,0 +1,9 @@
+const { Router } = require('express');
+const { LGUTimetable } = require('../controllers/LGUTimetable');
+const { Timetable } = require('../models');
+
+const router = Router();
+
+router.get('/', LGUTimetable.get);
+
+module.exports = router;
